fix: 의존성 순환 문제로 별도의 유저 검증 서비스 생성
