+++ src/main/java/com/sky7th/deliveryfood/security/service/AuthService.java
@@ -7,19 +7,13 @@ import com.sky7th.deliveryfood.security.refreshtoken.RefreshTokenService;
 import com.sky7th.deliveryfood.user.CustomUserDetails;
 import com.sky7th.deliveryfood.user.LoginResponseDto;
 import com.sky7th.deliveryfood.user.TokenRefreshRequestDto;
-import com.sky7th.deliveryfood.user.User;
 import com.sky7th.deliveryfood.user.UserContext;
-import com.sky7th.deliveryfood.user.UserRole;
-import com.sky7th.deliveryfood.user.member.domain.Member;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.LockedException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
@@ -29,24 +23,6 @@ public class AuthService {
   private final AuthenticationManager authenticationManager;
   private final RefreshTokenService refreshTokenService;
   private final JwtTokenProvider jwtTokenProvider;
-  private final BCryptPasswordEncoder bCryptPasswordEncoder;
-
-  public void validateUser(User user, String enteredPassword) {
-    if (!bCryptPasswordEncoder.matches(enteredPassword, user.getPassword())) {
-      throw new BadCredentialsException(user.getEmail());
-
-    } else if (!user.getActive()) {
-      throw new LockedException(user.getEmail());
-    }
-
-    if (user.getRole() == UserRole.ROLE_MEMBER) {
-      Member member = (Member) user;
-
-      if (!member.getEmailVerified()) {
-        user.setRole(UserRole.ROLE_GUEST_MEMBER);
-      }
-    }
-  }
 
   public CustomUserDetails authenticateUser(
       UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken) {
